requirements.txt
playwright
playwright-stealth
pandas
phonenumbers
gspread
google-auth
python-dotenv
